Fix "broken" test
This is tricky. There appears to be a flaw in the businesses logic for certain
item types. Some quality updates are calculated before decrementing the lifespan,
others are calculated after.

In the absence of a domain expert to consult, I’ve decided to regard this as a
bug in the current implementation.